+++ instrumentation/opentelemetry-instrumentation-django/src/opentelemetry/instrumentation/django/middleware.py
@@ -79,6 +79,7 @@ try:
         collect_request_attributes as asgi_collect_request_attributes,
         set_status_code,
     )
+
     _is_asgi_supported = True
 except ImportError:
     asgi_getter = None
@@ -252,7 +253,7 @@ class _DjangoMiddleware(MiddlewareMixin):
 
         if activation and span:
             if is_asgi_request:
-                set_status_code(request.META[self._environ_span_key], response.status_code)
+                set_status_code(span, response.status_code)
             else:
                 add_response_attributes(
                     span,